Refactor build system, grid renderer, and graphing
- Switched to explicit source listing and C++23 in CMake; added precompiled headers and improved dependency management
- Renamed and refactored SpatialRenderer to SpatialGridRenderer with encapsulated implementation and more efficient rendering
- Increased world/food population and grid size limits; made settings runtime-tunable
- Improved PopulationHistory with ring buffer eviction and event pruning; enhanced graphing with cached bands and flexible Y-axis fitting
- Modernized includes, added forward declarations, and improved code clarity throughout
- Fixed type mismatches, renderer calls, and event marker placement in graphs
- Overall, improves performance, scalability, and maintainability

+++ src/Protozoa/Protozoa.h
@@ -1,5 +1,8 @@
 #pragma once
-#include <SFML/Graphics.hpp>
+#include <SFML/Graphics/RenderWindow.hpp>
+#include <SFML/Graphics/Rect.hpp>
+#include <SFML/System/Vector2.hpp>
+
 #include "../settings.h"
 #include "../Utils/Graphics/Circle.h"
 #include "cell.h"
@@ -46,7 +49,8 @@ private:
 	std::vector<Spring> m_springs_{};	
 
 	// reproduction related parameters
-	FixedSpan<obj_idx> nearby_food_container{ FoodSettings::cell_max_capacity * 9 };
+	FixedSpan<obj_idx> nearby_food_container{
+	static_cast<std::uint8_t>(FoodSettings::cell_max_capacity * 9)};
 
 	unsigned stomach = 0;
 	float energy = ProtozoaSettings::initial_energy;
